+++ lib/klaviyo_api/resources/profile.rb
@@ -2,6 +2,8 @@
 
 module KlaviyoAPI
   class Profile < Base
+    extend KlaviyoAPI::Support::Countable
+
     self.prefix += 'v1/'
 
     self.element_name = 'person'