+++ lib/klaviyo_api.rb
@@ -21,4 +21,5 @@ require 'klaviyo_api/detailed_log_subscriber'
 require 'klaviyo_api/exceptions'
 require 'klaviyo_api/session'
 
+require 'klaviyo_api/collections'
 require 'klaviyo_api/resources'