+++ lib/klaviyo_api/resources/support/countable.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module KlaviyoAPI::Support
+  module Countable
+    def count(params = {})
+      all(params).total
+    end
+  end
+end