+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
  - Support `/v1/person/<id>`
  - Support `/v1/people`
+ - Support `.count` on Profile
 
 ## [1.2.0]